add interrupted tests
