+++ Makefile
@@ -45,7 +45,7 @@ src/curr.rs: $(XDR_FILES_LOCAL_CURR)
 	> $@
 	docker run -i --rm -v $$PWD:/wd -w /wd docker.io/library/ruby:latest /bin/bash -c '\
 		gem install specific_install -v 0.3.7 && \
-		gem specific_install https://github.com/stellar/xdrgen.git -b 11216c21 && \
+		gem specific_install https://github.com/stellar/xdrgen.git -b 3dd831ac && \
 		xdrgen --language rust --namespace curr --output src/ $^ \
 		'
 	rustfmt $@
@@ -54,7 +54,7 @@ src/next.rs: $(XDR_FILES_LOCAL_NEXT)
 	> $@
 	docker run -i --rm -v $$PWD:/wd -w /wd docker.io/library/ruby:latest /bin/bash -c '\
 		gem install specific_install -v 0.3.7 && \
-		gem specific_install https://github.com/stellar/xdrgen.git -b 11216c21 && \
+		gem specific_install https://github.com/stellar/xdrgen.git -b 3dd831ac && \
 		xdrgen --language rust --namespace next --output src/ $^ \
 		'
 	rustfmt $@
